Call stop() on websocket and recreate it
Reusing the connection segfaults,
at least if it's reused quickly after calling close().
Adding any kind of sleep fixes that sometimes,
but doing that is undesirable.
Calling stop() stops the eventloop
and renders the websocket instance useless,
so another one has to be created.
stop() also blocks until the eventloop exited, which is nice.